+++ source/rust/vndf/src/shared/game.rs
@@ -36,7 +36,7 @@ pub struct ManeuverData {
 #[derive(Clone, Copy, Debug, PartialEq, RustcDecodable, RustcEncodable)]
 pub struct Planet {
 	pub color: Color,
-	pub size : f32,
+	pub size : f64,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, RustcDecodable, RustcEncodable)]
